GH#28788: Isolate collector health and outcome state
