+++ .agents/scripts/knowledge_collector_schedule.py
@@ -25,13 +25,13 @@ from _knowledge_collector_model import (  # noqa: E402
     SCHEMA,
     Connection,
     _connection,
-    _health,
     _record,
     due_at,
     load_config,
     load_state,
-    plan,
 )
+from _knowledge_collector_health import _health, plan  # noqa: E402
+from _knowledge_collector_outcome import _apply_receipt, _record_failure  # noqa: E402
 from _knowledge_collector_runtime import (  # noqa: E402
     CollectorInterrupted,
     CollectorScheduleError,
@@ -47,8 +47,6 @@ DEFAULT_STATE = (
     / "knowledge"
     / "collector-health.json"
 )
-COUNT_KEYS = ("pages", "items", "bytes", "budget_units")
-OPTIONAL_RECEIPT_KEYS = (*COUNT_KEYS, "rate_reset_at", "coverage_status", "budget_stop")
 
 
 def build_command(connection: Connection) -> list[str]:
@@ -92,63 +90,6 @@ def _run_projection(connection: Connection, timeout: int) -> str:
         if result.returncode != 0:
             return "failed"
     return "complete"
-
-
-def _record_failure(record: dict[str, Any], connection: Connection, now: int) -> None:
-    failures = int(record.get("consecutive_failures", 0)) + 1
-    record.update(
-        {
-            "alert": failures >= connection.alert_after_failures,
-            "collector_outcome": "failed",
-            "consecutive_failures": failures,
-            "last_terminal_failure": now,
-            "status": "failed",
-        }
-    )
-
-
-def _apply_receipt(
-    record: dict[str, Any],
-    connection: Connection,
-    receipt: dict[str, Any],
-    now: int,
-    returncode: int,
-) -> tuple[int, bool]:
-    changed = receipt.pop("changed_count")
-    disposition = receipt.pop("disposition")
-    if returncode != 0 and disposition == "complete":
-        disposition = "failed"
-    projection_pending = connection.projection_root is not None and (
-        changed > 0 or record.get("projection_status") == "pending"
-    )
-    for key in OPTIONAL_RECEIPT_KEYS:
-        record.pop(key, None)
-    record.update(
-        {
-            "changed_count": changed,
-            "collector_outcome": disposition,
-            "projection_status": "pending" if projection_pending else "not-needed",
-            **receipt,
-        }
-    )
-    if disposition == "failed":
-        _record_failure(record, connection, now)
-    elif disposition == "deferred":
-        record.update({"last_deferred": now, "status": "pending"})
-    elif disposition == "partial":
-        record.update({"last_partial": now, "status": "partial"})
-    else:
-        record.update(
-            {
-                "alert": False,
-                "consecutive_failures": 0,
-                "last_success": now,
-                "status": "partial" if projection_pending else "complete",
-            }
-        )
-        if connection.event_token is not None:
-            record["event_token"] = connection.event_token
-    return changed, projection_pending
 
 
 @dataclass(frozen=True)
